user can apply a coupon discount code feature added

+++ app/Http/Requests/Admin/ProductUpdateRequest.php
@@ -29,7 +29,7 @@ class ProductUpdateRequest extends FormRequest
             'long_description' => ['required'],
             'price' => ['required', 'numeric'],
             'offer_price' => ['nullable', 'numeric'],
-            'quantity' => ['required', 'numeric'],
+            'quantity' => ['required', 'integer'],
             'sku' => ['nullable', 'max:255'],
             'seo_title' => ['nullable', 'max:255'],
             'seo_description' => ['nullable', 'max:500'],
